// frontend/src/core/api/generated/model/listMaterialMovementsParams.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Kızılkan Şantiye API
 * OpenAPI spec version: 1.0
 */
import type { ListMaterialMovementsDirection } from './listMaterialMovementsDirection'
import type { ListMaterialMovementsSort } from './listMaterialMovementsSort'
import type { ListMaterialMovementsStatus } from './listMaterialMovementsStatus'
import type { ListMaterialMovementsType } from './listMaterialMovementsType'

export type ListMaterialMovementsParams = {
  /**
   * @nullable
   */
  from?: string | null
  /**
   * @nullable
   */
  to?: string | null
  /**
   * @nullable
   */
  locationId?: string | null
  /**
   * @nullable
   */
  materialId?: string | null
  /**
   * @nullable
   */
  partyId?: string | null
  /**
   * @nullable
   */
  type?: ListMaterialMovementsType
  /**
   * @nullable
   */
  status?: ListMaterialMovementsStatus
  /**
   * @nullable
   */
  category?: string | null
  /**
   * @nullable
   */
  q?: string | null
  /**
   * @nullable
   */
  page?: number | null
  /**
   * @nullable
   */
  size?: number | null
  /**
   * @nullable
   */
  sort?: ListMaterialMovementsSort
  /**
   * @nullable
   */
  direction?: ListMaterialMovementsDirection
}